requirements.txt
streamlit
torch
opencv-python-headless
numpy
Pillow
segment-anything
